feat(filter): add a filter component
adds a new component called filter to handle the categories filtering and price filtering this
component will handle all the filters and display it in the home page it contain the filters
components and the buttons of the filters and display it in one file to be more readable and order

BREAKING CHANGE: changing the previous ProductFilter component to be CategoryFilter and extract the
buttons inside it to be another component to make it more readable and ordered

"fix #6"
