+++ src/components/Filter/Buttons/index.jsx
@@ -1,0 +1,19 @@
+export default function Buttons({ filters = [], handleFilterClick }) {
+  const FilterButton = ({ filter }) => (
+    <button
+      className={`filter-${filter}`}
+      onClick={() => handleFilterClick(filter)}
+    >
+      {filter}
+    </button>
+  )
+
+  return (
+    <div>
+      <button onClick={() => handleFilterClick("")}>All</button>
+      {filters.map((filter) => (
+        <FilterButton key={filter} filter={filter} />
+      ))}
+    </div>
+  )
+}
