+++ src/components/Filter/Price/index.jsx
@@ -1,0 +1,3 @@
+export default function PriceFilter() {
+  return <div>PriceFilter</div>
+}
