+++ src/components/Filter/index.jsx
@@ -1,0 +1,9 @@
+import Category from "./Category"
+
+export default function Filter() {
+  return (
+    <div>
+      <Category />
+    </div>
+  )
+}
